+++ frontend/toy/src/App.js
@@ -3,7 +3,6 @@ import { Routes, Route } from 'react-router-dom';
 import BoardDetail from './board/Article/BoardDetail';
 import Home from './Home';
 import UpdateArticle from './board/Article/UpdateArticle';
-import UpdateComment from './board/Comment/UpdateComment';
 
 function App() {
   return (
@@ -12,7 +11,6 @@ function App() {
         <Route path="/" element={<Home/>}/>
         <Route path="/article/:articleId" element={<BoardDetail/>}/>
         <Route path="/update/:articleId" element={<UpdateArticle/>}/>
-        <Route path="/update/comment/:commentId" element={<UpdateComment/>}/>
       </Routes>
     </>
   )
